parser/htmlparser/tests/mochitest/test_bug709083.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=709083
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 709083</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=709083">Mozilla Bug 709083</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<video muted>
<script type="application/javascript">

/** Test for Bug 709083 **/

ok(document.getElementsByTagName("video")[0].muted, "Should be muted already.");

</script>
</video>
</pre>
</body>
</html>
